Move pattern matchers to test support
